Разносит моки по отдельным файлам
