+++ src/components/reviews-list/reviews-list.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import {offers} from "../../mocks/forTest";
+import {offers} from "../../mocks/for-test/offers";
 import ReviewsList from "./reviews-list";
 
 it(`should render ReviewsList`, function () {
