+++ src/components/reviews/reviews.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import Reviews from "./reviews";
-import {offers} from "../../mocks/forTest";
+import {offers} from "../../mocks/for-test/offers";
 
 it(`should render Reviews`, function () {
   const tree = renderer
