+++ src/components/sort-item/sort-item.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import SortItem from "./sort-item";
-import {SortType} from "../../mocks/forTest";
+import {SortType} from "../../mocks/for-test/const";
 
 it(`should render SortItem`, function () {
   const tree = renderer
