+++ src/mocks/for-test/cities.js
@@ -1,0 +1,32 @@
+export const cities = [
+  {
+    title: `Paris`,
+    isActive: false,
+    coordinates: [48.856663, 2.351556]
+  },
+  {
+    title: `Cologne`,
+    isActive: false,
+    coordinates: [47.010209, 2.320878]
+  },
+  {
+    title: `Brussels`,
+    isActive: false,
+    coordinates: [50.851309, 4.351718]
+  },
+  {
+    title: `Amsterdam`,
+    isActive: false,
+    coordinates: [52.38333, 4.9]
+  },
+  {
+    title: `Hamburg`,
+    isActive: false,
+    coordinates: [53.552645, 9.966287]
+  },
+  {
+    title: `Dusseldorf`,
+    isActive: false,
+    coordinates: [51.230569, 6.787428]
+  },
+];
