+++ src/mocks/for-test/const.js
@@ -1,0 +1,8 @@
+export const SortType = {
+  POPULAR: `Popular`,
+  PRICE_UP: `Price: low to high`,
+  PRICE_DOWN: `Price: high to low`,
+  TOP_RATING: `Top rated first`
+};
+
+export const SORT_TYPES = Object.values(SortType);
