+++ src/mocks/for-test/offers.js
@@ -1,4 +1,4 @@
-import {TypePlace} from "../constants";
+import {TypePlace} from "../../constants";
 
 export const offers = [
   {
@@ -89,45 +89,3 @@ export const offers = [
     ]
   },
 ];
-
-export const cities = [
-  {
-    title: `Paris`,
-    isActive: false,
-    coordinates: [48.856663, 2.351556]
-  },
-  {
-    title: `Cologne`,
-    isActive: false,
-    coordinates: [47.010209, 2.320878]
-  },
-  {
-    title: `Brussels`,
-    isActive: false,
-    coordinates: [50.851309, 4.351718]
-  },
-  {
-    title: `Amsterdam`,
-    isActive: false,
-    coordinates: [52.38333, 4.9]
-  },
-  {
-    title: `Hamburg`,
-    isActive: false,
-    coordinates: [53.552645, 9.966287]
-  },
-  {
-    title: `Dusseldorf`,
-    isActive: false,
-    coordinates: [51.230569, 6.787428]
-  },
-];
-
-export const SortType = {
-  POPULAR: `Popular`,
-  PRICE_UP: `Price: low to high`,
-  PRICE_DOWN: `Price: high to low`,
-  TOP_RATING: `Top rated first`
-};
-
-export const SORT_TYPES = Object.values(SortType);
