+++ src/reducer.test.js
@@ -1,8 +1,10 @@
 import {reducer, ActionType, ActionCreator} from "./reducer";
 import {PageType} from "./constants";
-import {offers as testOffers, cities, SORT_TYPES, SortType} from "./mocks/forTest";
 import {extend, getOffersBySort} from "./utils";
 import offers from "./mocks/offers";
+import {SORT_TYPES, SortType} from "./mocks/for-test/const";
+import {offers as testOffers} from "./mocks/for-test/offers";
+import {cities} from "./mocks/for-test/cities";
 
 const someOffers = (activeCity) => offers.filter((offer) => offer.city === activeCity.title);
 const getCityList = (index) => cities.map((it) => {
